[RM-5896] Ability to sync custom rules into user defined family (#40)
Destination families may be provided in either the __rego_metadoc__
block within each rule, or on the command line using
`fugue sync rules --target-rule-family ...`.  If provided in the
metadoc block you may use either UUIDs or names, but if provided
on the command line only names are accepted.  The command line takes
priority over the metadoc and can therefore be used to override the
metadoc.

Note that if a family does not already exist you must create it before
attempting to add rules to it, otherwise an error will be emitted.

Example for the __rego_metadoc__ block:

```
__rego__metadoc__ := {
  "id": "FG_R00011",
  "title": "CloudFront viewer protocol policy should be set to https-only or redirect-to-https",
  "description": "CloudFront viewer protocol policy should be set to https-only or redirect-to-https. CloudFront connections should be encrypted during transmission over networks that can be accessed by malicious individuals. A CloudFront distribution should only use HTTPS or Redirect HTTP to HTTPS for communication between viewers and CloudFront.",
  "custom": {
    "severity": "Medium",
    "families": [
      "MyCustomFamily",
      "1172ca4f-6d31-4c46-a085-54ff73c6ed27"
    ]
  }
}
```

This would cause the rule to be inserted into both the family named
"MyCustomFamily" *and* the family with a UUID of
1172ca4f-6d31-4c46-a085-54ff73c6ed27.